Configure el debugger en CLion

+++ x64BareBones/run.sh
@@ -19,11 +19,13 @@ fi
 
 if [[ "$1" == "gdb" ]]; then
     qemu-system-x86_64 -s -S -hda Image/x64BareBonesImage.qcow2 -m 512 -d int $AUDIO_FLAGS -machine pcspk-audiodev=speaker > /dev/null 2>&1 &
-    QEMU_PID=$!
-    docker start Sistemas-Operativos-TP2
-    docker exec -it Sistemas-Operativos-TP2 bash -c "cd root && gdb"
-    docker stop Sistemas-Operativos-TP2
-    kill $QEMU_PID
-    else
+    if [[ "$2" == "docker" ]]; then
+      QEMU_PID=$!
+      docker start Sistemas-Operativos-TP2
+      docker exec -it Sistemas-Operativos-TP2 bash -c "cd root && gdb"
+      docker stop Sistemas-Operativos-TP2
+      kill $QEMU_PID
+    fi
+else
     qemu-system-x86_64 -hda Image/x64BareBonesImage.qcow2 -m 512 $AUDIO_FLAGS -machine pcspk-audiodev=speaker
 fi
